Exported nagivation to its own file
